refactor(ds): define interfaces in place
