+++ ui/design-system-core/src/components/Avatar/index.tsx
@@ -3,8 +3,30 @@ import { tw } from '@twind/core';
 
 import AvatarImage from './AvatarImage';
 
-import { IAvatarProps } from '../../interfaces/avatar.interface';
 import { getAvatarFromSeed } from '../../utils/get-avatar-from-seed';
+
+export type AvatarSrc = { url?: string; fallbackUrl?: string };
+
+export type AvatarSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
+
+export type AvatarBorderSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
+
+export type AvatarBorderColor = 'white' | 'darkerBlue' | 'accent';
+
+export interface IAvatarProps {
+  ethAddress?: string | null;
+  alt?: string;
+  publicImgPath?: string;
+  backgroundColor?: string;
+  src?: AvatarSrc;
+  size?: AvatarSize;
+  border?: AvatarBorderSize;
+  borderColor?: AvatarBorderColor;
+  faded?: boolean;
+  active?: boolean;
+  isClickable?: boolean;
+  onClick?: () => void;
+}
 
 export const avatarSizesMap = {
   xs: '6',
