--- ui/design-system-core/src/interfaces/appAvatar.interface.ts
@@ -1,6 +1,0 @@
-import { IntegrationTypes } from '@akashaorg/typings/ui';
-import { IAvatarProps } from './avatar.interface';
-
-export interface IAppAvatarProps extends IAvatarProps {
-  appType: IntegrationTypes;
-}
