--- ui/design-system-core/src/interfaces/basicCardBox.interface.ts
@@ -1,11 +1,0 @@
-export interface IBasicCardBox {
-  elevation?: 'sm' | 'md' | 'lg' | 'xl' | '2xl' | 'inner' | 'none';
-  dashedBorder?: boolean;
-  rootNodeRef?: React.Ref<HTMLDivElement>;
-  pad?: string;
-  margin?: string;
-  round?: string;
-  noBorder?: boolean;
-  noBorderRadius?: boolean;
-  onClick?: () => void;
-}
