--- ui/design-system-core/src/interfaces/basicpopover.interface.ts
@@ -1,9 +1,0 @@
-export type OverflowType = 'auto' | 'hidden' | 'scroll' | 'visible';
-
-export interface IBasicPopover {
-  children: React.ReactNode;
-  closePopover: () => void;
-  target: HTMLElement;
-  gap?: string;
-  overflow?: OverflowType | { horizontal?: OverflowType; vertical?: OverflowType };
-}
