--- ui/design-system-core/src/interfaces/stackedAvatar.interface.ts
@@ -1,10 +1,0 @@
-import { IProfileData } from '@akashaorg/typings/ui';
-import { AvatarSize } from './avatar.interface';
-
-export type UserDataType = { ethAddress: string; avatar?: IProfileData['avatar'] }[];
-
-export interface IStackedAvatarProps {
-  userData: UserDataType;
-  maxAvatars?: number;
-  size?: AvatarSize;
-}
